+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Clean up old step from README - It is no longer recommended to define footnotes in `WAGTAILADMIN_RICH_TEXT_EDITORS`
 - Add `footnotes` rich text feature automatically
 - Clean up `RichTextBlockWithFootnotes`
+- Make Footnotes unique on `page` <-> `uuid`
 
 ## 0.6.0